+++ index.js
@@ -17,3 +17,4 @@ module.exports.emitters = process.requireApi('lib/emitters/index.js');
 module.exports.grunt = process.requireApi('lib/grunt/index.js');
 module.exports.multipart = process.requireApi('lib/multipart/index.js');
 module.exports.passport = process.requireApi('lib/passport/index.js');
+module.exports.watcher = process.requireApi('lib/watcher/index.js');