+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 
 - require('@openveo/api').logger now adds a timestamp property to each log containing localized date and time
 - Improve code documentation by replacing Yuidoc by JSDoc
+- Add require('@openveo/api').watcher to listen to files system changes
+- ovWatch script has been added to execute an NPM script of files changes and can be launched using npx (see scripts/watch.js) for usage
 
 ## BUG FIXES